LayoutTests/css3/calc/getComputedStyle-margin-percentage.html

<!DOCTYPE html>
<script src="../../fast/js/resources/js-test-pre.js"></script>
<style>
#control { 
    margin-left: 33%;
    margin-top: 10%;
    margin-right: 45%;
    margin-bottom: 69%;
}
#calc { 
    margin-left: -webkit-calc(33%);
    margin-top: -webkit-calc(10%);
    margin-right: -webkit-calc(45%);
    margin-bottom: -webkit-calc(69%);
}
</style>
<div id="control"></div>
<div id="calc"></div>
<script>
    function test(prop) {
        shouldEvaluateTo("getComputedStyle(document.getElementById('calc'), null).getPropertyValue('" + prop + "')", "getComputedStyle(document.getElementById('control', null)).getPropertyValue('" + prop + "')");
    }
    
    description("Tests calling getPropertyValue on computed margin styles involving CSS calc");
    
    test('margin-left');
    test('margin-top');
    test('margin-right');
    test('margin-bottom');
</script>
<script src="../../fast/js/resources/js-test-post.js"></script>
